Fix signed/unsigned comparison mismatches in VelocityBuffer and PropKeys
VelocityBuffer::AdvanceFrame: cast mFrame to unsigned int for >= 2 comparison, achieving 100% match. FloatKeys::FloatAt: cast size() to int for < 3 comparison, improving from 88.6% to 89.1%.

+++ src/system/rndobj/PropKeys.cpp
@@ -341,7 +341,7 @@ int FloatKeys::FloatAt(float frame, float &fl) {
         Interp(prev->value, next->value, ref, fl);
         break;
     case kSpline:
-        if (size() < 3 || prev == next) {
+        if ((int)size() < 3 || prev == next) {
             Interp(prev->value, next->value, ref, fl);
         } else {
             float points[4];
